Automatically forward declare classes for the rcpp stubs
Will break in a number of cases (classes defined within classes,
template classes, etc) but should work for now at least.  This adds
two more keys to the class part of the yml: 'is_struct', which
needs to be set to "True" if the exported thing is actually a struct
rather than a class, and "namespace", which will be needed if the
object is not in the root namespace and can't be easily inferred.
Otherwise we assume all qualifiers are namespace.