Initial support for list object export
This allows objects to be easily serialised as a list.  It's
not pretty, but I need this for some classes that are fundamentally
just heterogeneous lists of data, with no methods.  With
this export there is no reference semantics, and each exported
field is simply copied between an R list (Rcpp::List) and the
C++ object.